# Conflicts:
#	docker-compose.yml
#	go.sum
#	internal/config/flags.go
#	internal/config/options.go

+++ internal/entity/file.go
@@ -7,6 +7,8 @@ import (
 	"strings"
 	"time"
 
+	"github.com/dustin/go-humanize/english"
+
 	"github.com/gosimple/slug"
 	"github.com/jinzhu/gorm"
 	"github.com/ulule/deepcopier"
@@ -200,22 +202,26 @@ func (m File) Missing() bool {
 	return m.FileMissing || m.DeletedAt != nil
 }
 
-// DeletePermanently permanently deletes a file from the index.
+// DeletePermanently permanently removes a file from the index.
 func (m *File) DeletePermanently() error {
+	if m.ID < 1 || m.FileUID == "" {
+		return fmt.Errorf("invalid file id %d / uid %s", m.ID, txt.Quote(m.FileUID))
+	}
+
 	if err := UnscopedDb().Delete(Marker{}, "file_uid = ?", m.FileUID).Error; err != nil {
-		log.Errorf("file: %s (delete markers)", err)
+		log.Errorf("file %s: %s while removing markers", txt.Quote(m.FileUID), err)
 	}
 
 	if err := UnscopedDb().Delete(FileShare{}, "file_id = ?", m.ID).Error; err != nil {
-		log.Errorf("file: %s (delete shares)", err)
+		log.Errorf("file %s: %s while removing share info", txt.Quote(m.FileUID), err)
 	}
 
 	if err := UnscopedDb().Delete(FileSync{}, "file_id = ?", m.ID).Error; err != nil {
-		log.Errorf("file: %s (delete sync)", err)
+		log.Errorf("file %s: %s while removing remote sync info", txt.Quote(m.FileUID), err)
 	}
 
 	if err := m.ReplaceHash(""); err != nil {
-		log.Errorf("file: %s (delete previews)", err)
+		log.Errorf("file %s: %s while removing covers", txt.Quote(m.FileUID), err)
 	}
 
 	return UnscopedDb().Delete(m).Error
@@ -230,9 +236,11 @@ func (m *File) ReplaceHash(newHash string) error {
 
 	// Log values.
 	if m.FileHash != "" && newHash == "" {
-		log.Tracef("file: removing hash %s", txt.Quote(m.FileHash))
+		log.Tracef("file %s: removing hash %s", txt.Quote(m.FileUID), txt.Quote(m.FileHash))
 	} else if m.FileHash != "" && newHash != "" {
-		log.Tracef("file: hash %s changed to %s", txt.Quote(m.FileHash), txt.Quote(newHash))
+		log.Tracef("file %s: hash %s changed to %s", txt.Quote(m.FileUID), txt.Quote(m.FileHash), txt.Quote(newHash))
+		// Reset error when hash changes.
+		m.FileError = ""
 	}
 
 	// Set file hash to new value.
@@ -255,10 +263,8 @@ func (m *File) ReplaceHash(newHash string) error {
 
 		if res := UnscopedDb().Model(entity).Where("thumb = ?", oldHash).UpdateColumn("thumb", newHash); res.Error != nil {
 			return res.Error
-		} else if res.RowsAffected == 1 {
-			log.Infof("%s: updated %d preview [%s]", name, res.RowsAffected, time.Since(start))
-		} else if res.RowsAffected > 1 {
-			log.Infof("%s: updated %d previews [%s]", name, res.RowsAffected, time.Since(start))
+		} else if res.RowsAffected > 0 {
+			log.Infof("%s: updated %s [%s]", name, english.Plural(int(res.RowsAffected), "cover", "covers"), time.Since(start))
 		}
 	}
 
@@ -267,6 +273,10 @@ func (m *File) ReplaceHash(newHash string) error {
 
 // Delete deletes the entity from the database.
 func (m *File) Delete(permanently bool) error {
+	if m.ID < 1 || m.FileUID == "" {
+		return fmt.Errorf("invalid file id %d / uid %s", m.ID, txt.Quote(m.FileUID))
+	}
+
 	if permanently {
 		return m.DeletePermanently()
 	}
@@ -306,16 +316,16 @@ func (m *File) AllFilesMissing() bool {
 // Create inserts a new row to the database.
 func (m *File) Create() error {
 	if m.PhotoID == 0 {
-		return fmt.Errorf("file: photo id must not be empty (create)")
+		return fmt.Errorf("file: can't create file with empty photo id")
 	}
 
 	if err := UnscopedDb().Create(m).Error; err != nil {
-		log.Errorf("file: %s (create)", err)
+		log.Errorf("file: %s while saving", err)
 		return err
 	}
 
 	if _, err := m.SaveMarkers(); err != nil {
-		log.Errorf("file: %s (create markers for %s)", err, m.FileUID)
+		log.Errorf("file %s: %s while saving markers", txt.Quote(m.FileUID), err)
 		return err
 	}
 
@@ -334,16 +344,16 @@ func (m *File) ResolvePrimary() error {
 // Save stores the file in the database.
 func (m *File) Save() error {
 	if m.PhotoID == 0 {
-		return fmt.Errorf("file: photo id must not be empty (save %s)", m.FileUID)
+		return fmt.Errorf("file %s: can't save file with empty photo id", m.FileUID)
 	}
 
 	if err := UnscopedDb().Save(m).Error; err != nil {
-		log.Errorf("file: %s (save %s)", err, m.FileUID)
+		log.Errorf("file %s: %s while saving", txt.Quote(m.FileUID), err)
 		return err
 	}
 
 	if _, err := m.SaveMarkers(); err != nil {
-		log.Errorf("file: %s (save markers for %s)", err, m.FileUID)
+		log.Errorf("file %s: %s while saving markers", txt.Quote(m.FileUID), err)
 		return err
 	}
 
@@ -373,7 +383,7 @@ func (m *File) Updates(values interface{}) error {
 
 // Rename updates the name and path of this file.
 func (m *File) Rename(fileName, rootName, filePath, fileBase string) error {
-	log.Debugf("file: renaming %s to %s", txt.Quote(m.FileName), txt.Quote(fileName))
+	log.Debugf("file %s: renaming %s to %s", txt.Quote(m.FileUID), txt.Quote(m.FileName), txt.Quote(fileName))
 
 	// Update database row.
 	if err := m.Updates(map[string]interface{}{
@@ -417,7 +427,7 @@ func (m *File) Undelete() error {
 		return err
 	}
 
-	log.Debugf("file: removed missing flag from %s", txt.Quote(m.FileName))
+	log.Debugf("file %s: removed missing flag from %s", txt.Quote(m.FileUID), txt.Quote(m.FileName))
 
 	m.FileMissing = false
 	m.DeletedAt = nil
@@ -480,8 +490,8 @@ func (m *File) AddFaces(faces face.Faces) {
 
 // AddFace adds a face marker to the file.
 func (m *File) AddFace(f face.Face, subjUID string) {
-	// Only add faces with embedding, so that they can be clustered.
-	if len(f.Embeddings) != 1 {
+	// Only add faces with exactly one embedding so that they can be compared and clustered.
+	if !f.Embeddings.One() {
 		return
 	}
 
@@ -495,7 +505,7 @@ func (m *File) AddFace(f face.Face, subjUID string) {
 
 	// Append marker if it doesn't conflict with existing marker.
 	if markers := m.Markers(); !markers.Contains(*marker) {
-		markers.Append(*marker)
+		markers.AppendWithEmbedding(*marker)
 	}
 }
 
@@ -536,7 +546,7 @@ func (m *File) Markers() *Markers {
 	} else if m.FileUID == "" {
 		m.markers = &Markers{}
 	} else if res, err := FindMarkers(m.FileUID); err != nil {
-		log.Warnf("file: %s (load markers)", err)
+		log.Warnf("file %s: %s while loading markers", txt.Quote(m.FileUID), err)
 		m.markers = &Markers{}
 	} else {
 		m.markers = &res
